Refactor algorithms to separate file
